nrf51822: Remove local flasher
This patch removes the local flasher for the nRF51822 platform. This
flasher requires an old version of the J-Link software and adds an
dependence in the Makefile.platform.

Now it is recommended to use a new tool called `nrftool` to program the
nRF51822. This tool is a python package and can be installed using pip:

	$ [sudo] pip install nrftool

Or it can be downloaded at https://pypi.python.org/pypi/nrftool