bci: Add helper bci_ad_put()
This patch adds a helper function to append AD data types to the given
buffer. This function returns the number of octets written in the buffer, or
a negative value (-errno) if a wrong data type is informed. At the moment,
overflow is not handled, the caller must check if there are space enough to
insert the wanted AD types data.

Usage example:
size = bci_ad_put(buffer,
		BCI_AD_NAME_COMPLETE, name,
		BCI_AD_TX_POWER, tx,
		BCI_AD_INVALID);